Simplify Safari IPA generation scripts into a single consolidated file
